+++ forge/src/main/java/cc/cassian/clickthrough/helpers/forge/ModHelpersImpl.java
@@ -1,6 +1,6 @@
 package cc.cassian.clickthrough.helpers.forge;
 
-
+import cc.cassian.clickthrough.ArchitecturyImpl;
 import net.minecraft.block.BlockState;
 //? if >1.20 {
 import net.minecraft.registry.tag.BlockTags;
@@ -10,6 +10,9 @@ import net.minecraft.registry.tag.BlockTags;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.fml.ModList;
 
+import static cc.cassian.clickthrough.ClickThrough.*;
+import static cc.cassian.clickthrough.helpers.ModHelpers.architecturyInstalled;
+
 public class ModHelpersImpl {
     public static boolean clothConfigInstalled() {
         return ModList.get().isLoaded("cloth_config");
@@ -17,6 +20,15 @@ public class ModHelpersImpl {
 
     public static boolean isLoaded(String mod) {
         return ModList.get().isLoaded(mod);
+    }
+
+    public static void registerKeybind() {
+        if (architecturyInstalled()) {
+            ArchitecturyImpl.load();
+        }
+        else {
+            LOGGER.info("ClickThrough Plus running without Architectury. Keybinds are not avaialble!");
+        }
     }
 
     public static boolean isTaggedAsContainer(BlockState state) {