+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Empty Fast Item Frames can now be clicked through.
+- Fabric no longer uses Architectury API for keybinds.
 
 ## [3.4.1] - 2025-04-16